feat(inventory): allow fractional stock for products and adjustments
Start the Phase 1 DB-050 integer→real migration with the slice that
unblocks ferreterías (2.5 m cable) and supermarkets (0.75 kg produce):
product.stock/minStock plus the inventory_movements ledger now accept
real values end-to-end.

Schema:
- packages/server/src/db/schema.ts: products.stock, products.minStock,
  inventory_movements.quantity, .previousStock, .newStock are now real
- packages/server/src/db/index.ts: raw DDL bootstrap matches. SQLite
  type affinity lets existing INTEGER columns keep accepting the same
  numeric storage class, so no migration script is required

Validators:
- trpc/schemas/products.ts: drop .int() from create/update stock +
  minStock on four call sites
- trpc/schemas/inventory.ts: drop .int() from createMovementInput
  (tighten to .positive()) and adjustStockInput.newStock

Tests (+2 end-to-end round-trip cases, all through the tRPC caller):
- products.test.ts: create → getById → update → list preserves
  2.5 / 0.25 / 0.75 / 0.1 without rounding
- inventory.test.ts: adjustStock from 5 → 2.5 stores the magnitude
  (2.5) on inventory_movements and the new stock (2.5) on the product

UI (surgical fix so the browser number validator accepts decimals —
it defaults to step=1 which rejects non-integers):
- features/products/ProductFormModal.tsx: step='any' on stock +
  minStock inputs
- features/inventory/InventoryAdjustmentModal.tsx: step='any' on the
  new-stock target input

Out of scope for this slice: sale_items, purchase_items, order_items,
initial_inventory still carry integer quantity columns and will be
migrated in a follow-up. The roadmap's DB-050 ticket is now partially
landed with clear end-to-end tests pinning the working subset.

+++ packages/server/src/trpc/schemas/products.ts
@@ -66,8 +66,10 @@ export const createProductInput = z
     providerId: z.string().nullable().optional(),
     locationId: z.string().nullable().optional(),
     initialCost: z.number().min(0, 'Initial cost must be non-negative').default(0),
-    stock: z.number().int().min(0).default(0),
-    minStock: z.number().int().min(0).default(0),
+    // Phase 1 DB-050: stock accepts real numbers so ferreterías (2.5 m cable)
+    // and supermarkets (0.75 kg produce) can track fractional quantities.
+    stock: z.number().min(0).default(0),
+    minStock: z.number().min(0).default(0),
     isActive: z.boolean().default(true),
     barcode: z.string().nullable().optional(),
     imageUrl: z.string().nullable().optional(),
@@ -109,8 +111,9 @@ export const updateProductInput = z
     providerId: z.string().nullable().optional(),
     locationId: z.string().nullable().optional(),
     initialCost: z.number().min(0).optional(),
-    stock: z.number().int().min(0).optional(),
-    minStock: z.number().int().min(0).optional(),
+    // Phase 1 DB-050: see createProductInput above.
+    stock: z.number().min(0).optional(),
+    minStock: z.number().min(0).optional(),
     isActive: z.boolean().optional(),
     barcode: z.string().nullable().optional(),
     imageUrl: z.string().nullable().optional(),
